requirements.txt
streamlit==1.24.1
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.4.0
requests==2.31.0
 